Create FindMountainOrNot.java
Write a Java program that checks if a given array of integers forms a "mountain". A mountain array is defined as an array that firstly increases to a peak value and then decreases. For example, [1, 3, 5, 4, 2] is a mountain array because it increases from 1 to 5 and then decreases from 5 to 2.
Your program should implement a method isMountainArray that takes an integer array as input and returns true if the array is a mountain, and false otherwise.