Create DivisibleOrNotAfterRemovingOccur.java
Get two integers as input. One is a number and the other is a key. Remove all occurrences of the second input (key) from the first number and display the result. Then, check if the result is divisible by the second input (key) and display “Divisible” if it is divisible, or else display “Not Divisible”.